add Todo component, modified classes for Todos
