+++ src/components/Footer/Footer.jsx
@@ -1,8 +1,6 @@
-import React from 'react'
-import styles from './footer.module.css'
+import React from "react";
+import styles from "./footer.module.css";
 
 export default function Footer() {
-  return (
-    <div className={styles.footer}>Footer</div>
-  )
+  return <div className={styles.footer}>Footer</div>;
 }
