+++ src/components/Footer/footer.module.css
@@ -7,5 +7,4 @@
   line-height: 17px;
   text-align: center;
   color: #a9a9a9;
-
 }
