+++ src/components/Header/header.module.css
@@ -1,11 +1,11 @@
 .header {
-flex: 0 0 auto;
-margin-top: 32px;
-font-family: 'Raleway';
-font-style: normal;
-font-weight: 700;
-font-size: 36px;
-line-height: 42px;
-text-align: center;
-letter-spacing: -0.045em;
+  flex: 0 0 auto;
+  margin-top: 32px;
+  font-family: "Raleway";
+  font-style: normal;
+  font-weight: 700;
+  font-size: 36px;
+  line-height: 42px;
+  text-align: center;
+  letter-spacing: -0.045em;
 }
