+++ src/components/Main/Todos/Todo/todo.module.css
@@ -1,0 +1,30 @@
+.wrapperTodo {
+  margin-bottom: 28px;
+}
+
+.wrapperTodo:hover {
+  cursor: pointer;
+  transform: scale(1.01);
+}
+
+.checkbox {
+  width: 24px;
+  height: 24px;
+  box-sizing: border-box;
+  background: #2f80ed;
+  border-radius: 4px;
+}
+
+.textTodo {
+  margin-left: 7px;
+  font-family: "Montserrat";
+  font-style: normal;
+  font-weight: 500;
+  font-size: 18px;
+  line-height: 22px;
+  vertical-align: super;
+}
+
+.stroke {
+  text-decoration: line-through;
+}
