+++ src/components/Main/Todos/todos.module.css
@@ -1,23 +1,3 @@
 .wrapperTodo {
   margin-bottom: 28px;
 }
-
-.checkbox {
-  width: 24px;
-  height: 24px;
-  box-sizing: border-box;
-  background: #2f80ed;
-  border-radius: 4px;
-}
-.textTodo {
-  margin-left: 7px;
-  font-family: "Montserrat";
-  font-style: normal;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 22px;
-  vertical-align: super;
-}
-.stroke {
-  text-decoration: line-through;
-}
